make gdutils a package
