+++ pyproject.toml
@@ -1,0 +1,3 @@
+[build-system]
+requires = ["setuptools>=41.2", "setuptools_scm", "wheel"]
+build-backend = "setuptools.build_meta"
