+++ requirements-dev.txt
@@ -1,0 +1,22 @@
+black
+check-manifest
+doctr
+flake8
+flake8-builtins
+flake8-comprehensions
+flake8-mutable
+flake8-print
+interrogate
+isort
+nbsphinx
+pre-commit
+pylint
+pytest
+pytest-cov
+pytest-flake8
+pytest-xdist
+recommonmark
+setuptools_scm
+sphinx
+twine
+wheel
