+++ setup.cfg
@@ -1,0 +1,43 @@
+[metadata]
+name = gdutils
+description = Glider DAC Utilities
+author = John Kerfoot
+url = https://github.com/kerfoot/gdutils
+long_description_content_type = text/markdown
+long_description = file: README.md
+license = GPL-3.0
+license_file = LICENSE
+classifiers =
+    Development Status :: 5 - Production/Stable
+    Intended Audience :: Science/Research
+    Operating System :: OS Independent
+    License :: OSI Approved :: BSD License
+    Programming Language :: Python
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Topic :: Scientific/Engineering
+
+[options]
+zip_safe = True
+install_requires =
+    erddapy
+    seaborn
+python_requires = >=3.6
+packages = find:
+
+[sdist]
+formats = gztar
+
+[check-manifest]
+ignore =
+    *.yml
+    .coveragerc
+    docs
+    docs/*
+    *.enc
+    notebooks
+    notebooks/*
+    tests
+    tests/*
