+++ tests/test_import.py
@@ -1,0 +1,6 @@
+from gdutils import GdacClient
+
+
+
+def test_class():
+    return GdacClient()
